Moved io.github.skyblockcore.dungeons.DungeonUtils to io.github.skyblockcore.Dungeons, as per Unex's suggestion.
