+++ src/main/java/io/github/skyblockcore/event/dungeons/DungeonStartedCallback.java
@@ -1,6 +1,6 @@
 package io.github.skyblockcore.event.dungeons;
 
-import io.github.skyblockcore.dungeons.DungeonUtils;
+import io.github.skyblockcore.Dungeons;
 import net.fabricmc.fabric.api.event.Event;
 import net.fabricmc.fabric.api.event.EventFactory;
 import net.minecraft.util.ActionResult;
@@ -16,5 +16,5 @@ public interface DungeonStartedCallback {
                 return ActionResult.PASS;
             });
 
-    ActionResult interact(DungeonUtils.DUNGEON_CLASSES dungeonClass);
+    ActionResult interact(Dungeons.DUNGEON_CLASSES dungeonClass);
 }
